Reorganize classes into different files
